No more flask
In the end, as expected, I couldn't really get flask to work out with the prototype, which means that only the interactivity of the website is not functioning. Tomorrow I put in all the documentation that I have kept on a separate google document. But right now I go relax.